+++ src/explorer/embed.go
@@ -1,1 +1,6 @@
 package explorer
+
+import "embed"
+
+//go:embed static/*
+var StaticFiles embed.FS